move CDT struct/enum defs to separate module
